Allow VIM to evacuate instances from powered off host
When an active AIO-DX controller is powered down, maintenance sends
the VIM a host state change to the offline state and then to the
power-off state. The VIM may not attempt to evacuate instances
from the controller right away, because it is waiting for nova to
indicate the hypervisor on the newly active controller is
available. If the VIM doesn't attempt the evacuates until after
the power-off state is reached, the evacuates will not be allowed,
because the VIM only allows evacuates if the host is in the offline
state. The fix is to allow the VIM to evacuate instances if the
host is in either the offline or power-off states.

Change-Id: I436c0fb3107219445694e2aee25e758c3c08080a