Fix issue with stale alarm
When the VIM attempts to clear an alarm, but the request to FM
fails, the VIM is leaving the alarm in its internal list of raised
alarms. The next time the VIM's alarm audit runs, it will see that
the alarm is not raised and raise it. The alarm will then be
stuck.

The fix is to always remove the alarm from the list - even if FM
fails to clear the alarm. The VIM's alarm audit will then clear the
alarm because it will find an alarm in FM that is not in the VIM's
alarm list.

Change-Id: I011d17c00364601400cbd770ed31077e364f2c9f