Update VIM to use nova API version 2.53
Updating the VIM to use version 2.53 (Pike) of the nova API. The VIM
was previously using version 2.38 (Newton). The main changes are:
- Stop using os-hosts (no longer exists). Fortunately we weren't
  actually using this code.
- Use new format for os-services (server uuid now required in URL and
  body changed).
- The GET servers now returns a flavor in the response (instead of the
  flavor uuid). The VIM was previously using the uuid to look up the
  flavor in the instance_type table. Now the VIM must access the
  flavor information embedded in each instance.

Note that the server notifications from nova still include the flavor
uuid (we need to switch to the newer versioned notifications to get
the flavor). I changed our notification handling to ignore the flavor
uuid as it is no longer of any use to us.

The main reason for this change is to allow flavors to be modified or
deleted even if in use by a running instance. This brings us in line
with upstream behaviour and will allow us to stop modifying this
behaviour in nova.

Change-Id: I65fab6843174d59782aa7366f2e0cc646cebafb7
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>