Prevent patch strategy creation when AIO controller locked
In AIO-SX, we were allowing a patch orchestration strategy to be
created before the controller host had been unlocked. This causes
a reboot step to be added to the strategy, which doesn't work on
AIO-SX. The solution is to prevent a patch orchestration strategy
from being created when an AIO-SX host is locked - this was an
oversight in the original implementation.

Story: 2002863
Task: 22812

Change-Id: I4305fa91473e583b5c8e4b102e5f1678c33f912b
Signed-off-by: Don Penney <don.penney@windriver.com>