stx-nfv: API ref doc content added.
I added the api-ref-nfv-vim-v1.rst file, which is content for
the API reference manual to the api-ref/source dir. This
represents the converted old-style files to the newer
OpenStack supported RST files.

I updated the index.rst file to include the new .rst file
so that the api-ref document can build.

Change-Id: I8848b541a9983e63c48cbbc60fbdd42595c81655
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>