[WIP]ensure string instance_name "null-terminated"
strncpy is used to copy instance_name from source string to dest here.
As for strncpy does not guarantee null terminated of dest string, so
we need to do it manually.

changes: we bounded length of string instance_name to ensure it is
"null-terminated".

Test Case:
  1. one controller + one compute (virtual)
     success deploy and create 2 vm instance (one of the instance
name is longer than INSTANCE_NAME_SIZE 20)

Closes-Bug: 1794704

Change-Id: Ifb97e3dc1b59ebdc23cda73731fb02dc342d0520