Do not apply NoExecute taint to AIO hosts when locked
When a kubernetes host is locked, the VIM applies the NoExecute
taint to the node which causes pods to be evicted. However, in
the AIO simplex case, there is nowhere else for the pods to go
and when the host is unlocked, pods cannot start until the VIM
removes the NoExecute taint. The problem is that when the
OpenStack application is configured, the VIM requires the
rabbitmq pod to be running before the VIM can start, causing a
chicken and egg scenario. For now, we will avoid applying the
NoExecute taint on AIO hosts when they are locked, to avoid
this scenario.

Story: 2003910
Task: 27852

Change-Id: Icd7c00b6f7fba9cdbf23b60e265e90e9271d5243
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>