Mtce: Fix guestAgent inventory query
Maintenance was recently modified to support

 - receiving Ready Events and Command Responses and
 - sending Ready Events, Commands and Inventory

from/to surrogate daemons over its floating IP rather
than the LO interface in support of surrogate daemons
that run on the inactive controller or any hosts in
the system for that matter.

Unfortunately the guestAgent was not modified to send
its Ready Event or Receive commands and inventory in
the same way.

This update fixes that by binding its Inventory/Command
receiver and Ready event transmitter to the active
controllers floating IP.

With that change the guestAgent once again gets its
Inventory/Commands on its process startup as well as
on the restart of the mtcAgent.

Change-Id: I0b07596501f45d28ca0d6c82293d0b904a5e39fb
Closes-Bug: 1804579
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>