Replace iter methods with python3 compatable calls
Replace iteritems with items
Replace iterkeys with keys
Replace itervalues with values

Special Note:
In Python 2 those are copies of the elements.
In Python3 those are references to the dictionary elements.
The openstack python2to3 documents suggests wrapping the calls with
a list if the dict is expected to change while iterating.

nfv_scenario_tests (non guest) all pass.

Story: 2004241
Task: 28070
Change-Id: Iddb69b83f0be2f8ec240f7708d2c6d0dc675848c
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>