Do not apply NoExecute taint to hosts
The VIM is currently appying the NoExecute taint to controller
and compute hosts when they are disabled, in order to cause
the pods on those hosts to be evicted. This is causing
problems when configuring a non-AIO system because the
VIM goes through disable/enable cycles while the pods are
coming up on these hosts. This causes the NoExecute taint
to be applied/removed and prevents the OpenStack pods from
ever coming up successfully.

For now, the VIM will not apply the NoExecute taint. We will
revisit this later.

Change-Id: Icc8cece900c69b8d2620023a00ac906359d59d84
Story: 2003910
Task: 26796
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>