Fix relative import warnings from pylint
W0403 relative-import is no longer suppressed by pylint.rc

Note:
six.moves.urllib.request was considered a relative import
from six.moves import urllib is considered valid.

The code is updated to satisfy pylint and pep8.

The unit tests still use relative imports, but the codes
are suppressed on a per line basis.  This was done because the
unit tests are located in a module with '-' characters which
cause import issues if imported through their full path names.

Story: 2004515
Task: 28575
Change-Id: I3f576f3990352e9456fe3ee5d3a80cb1307f31e0
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>