Enable python3 unit testing in zuul for nfv
This does not ensure nfv is python3 compatable, but
it does ensure that the current unit tests and code covered
by those tests work in python3.

The shell.py changes are due to a change in default values in argparse.

The instance director change is because python3 uses a reference
rather than a copy, for its dict keys.

Story: 2003427
Task: 28818
Change-Id: Iaccc0ab5fc4e30b41df108f817612abad8ec612c
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>