Remove use of storage extra spec from VIM
The setting of the storage extra spec on each flavor was
removed from nova recently. However, the VIM still relies
on this extra spec being set and this results in the VIM
not making proper decisions on whether to live migrate,
cold migrate or evacuate instances during various system
events (e.g. a host lock).

The solution is to remove the storage extra spec from the
VIM. Live migration is now supported for instances,
regardless of the storage type. The cold migrate and
evacuate decisions will assume local storage is being
used until support for remote storage has been added to
the containers configuration.

Change-Id: I42d2d755c5cc61acb8e23bc278cd920a3b782562
Story: 2004386
Task: 28911
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>