Configure VIM nova listener queue as HA
In the pre-kubernetes configuration, we had a single instance of
rabbitmq running, so the VIM configured a single non-HA queue
to listen for nova notifications. In the kubernetes configuration
there are two instances of rabbitmq running. The nova notification
queue is configured to be HA through a policy that matches names
starting with "notifications.". However, the VIM listener queue
is named nfvi_nova_listener_queue, so it does not match the
policy and is not configured to be HA. That means that when the
home rabbitmq node goes away (e.g. due to a controller reboot),
the VIM listener queue is no longer available, which causes
failures in the VIM.

The solution is to rename the VIM queue to:
notifications.nfvi_nova_listener_queue
This will ensure that it matches the nova policy and is configured
with the same HA parameters as the nova notification queue.

Change-Id: I958aadfa2da8f4aba9d18631ee0808fe97c42032
story: 2002843
task: 29380
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>