Fix compute host delete with no kubernetes node
When a compute host is deleted, the VIM attempts to delete the
node from kubernetes. However, if a compute host was powered
on and the personality was set, but the host was never installed,
no kubernetes node will exist for the host. When the VIM attempts
to delete the node from kubernetes, the API returns a failure
and the VIM doesn't handle that properly, so it continues to
attempt to delete the node over and over.

The fix is to handle the NOT_FOUND response from the delete API
and treat that as a success.

Change-Id: If9ca8baae4252a46ed7a514b2e91586436262688
Story: 2002843
Task: 29496
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>