VIF model removal from nfvi
As part of the StarlingX strategy to align to openstack master,
the vif-model extension will no longer be available after the
rebase to Stein.

The NFVI/VIM code currently references the vif-model to identify
PCI-PT/SRIOV nics in order to prevent an instance with these
interfaces from live-migrating.

This commit refactors the check to make a call into Neutron and
use the vnic-type instead of the vif-model.

Change-Id: Id1526886f1a701eec71fa66170e3c7f546b82994
Story: 2005065
Task: 29633
Signed-off-by: Steven Webster <steven.webster@windriver.com>