Update py35 jobs to py37, run on Xenial
The upstream OpenStack default node type changed yesterday
from Ubuntu Xenial to Bionic. We have a number of jobs pegged
directly to py35 which is no longer available in Bionic.

We can either pin the py35 jobs to Xenial nodes or update them
to py36 or py37.  OpenStack is supporting both of those for the
Stein release, since StarlingX does not have any position on
Python 3 versions (yet!) I am suggesting we move straight to py37
and in the case of issues now make these jobs non-voting until
any incompatibilities are resolved.

Due to issues with psycopg2 we will continue to run this on Xenial
for now.

Change-Id: Ibb0260d29e48447d770f76211ad4af9b736d7907
Signed-off-by: Dean Troyer <dtroyer@gmail.com>