Ignore host audits with potentially outdated info
There are scenarios (e.g. host lock) where the VIM and the
maintenance component process aspects of the operation in
parallel. The VIM also periodically audits the host state
from maintenance. This opens up a window where the VIM can
get outdated state information from maintenance and then
reverse an operation that is in progress.

The solution is to update the VIM to ignore any audit
responses for hosts that have recently changed state. The
host audit is a non-critical audit that runs approximately
once a minute to ensure the VIM's state matches the
maintenance state and there will be no harm in ignoring
audits that occur shortly after a state change.

As added robustness, the VIM is also updated to apply the
a NoExecute taint to locked hosts when the host is
disabled outside an actual lock operation. This will
catch any cases where the host flips between the enabled
and disabled states while it is locked (e.g. due to the
scenario above).

Change-Id: I228a72956a4057c768c7cea12b63d967233e556d
Closes-Bug: 1820105
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>