Neutron Host Management Robustness
If a network agent goes down in the middle of processing
a reschedule/rebalance operation, abort the current operation
so as not to schedule resources onto a down agent.  This is
necessary as the reschedule/rebalance operations make take a
fair amount of time, and users may lock hosts (bringing
down the network agents) at any time.

Change-Id: Ie933b517e4e4cea3beb7d00e3c65d2f538f14e5c
Story: 2003857
Task: 30500
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>